Example: update to new module structure
