+++ modules/example/src/processors/simpleraycaster.cpp
